Moved "cleaning docs repo" script to other file.
